Implement role-based authentication and user creation
Changes in this commit primarily focus on enhancing the authentication mechanism by introducing role-based authentication, revamping the 'User' entity, and providing the functionality for user creation. The concept of roles with 'admin' and 'user' roles was introduced and tied to the 'User' entity instantiated from the repository. The 'AuthService' was also renamed to 'AuthenticationService' to better reflect its role, and methods were added for logging in and creating users. Also, to persist 'User' and 'Role' entities, 'JpaRepository' was used with H2 Database integrating into the project for data persistence during runtime.